Command Pattern
Added an example of the command pattern

- WASD to move (each press is one "move")
- Z to undo
- R to replay all moves (minus undos)

While it may seem  underwhelming, it is also currently under featured.
It should contain KeyDowns to start a Command, and KeyUp to end them, as
well as timestamps for these, and multi command functionality, for
starters.

That said, I think this will be useful to implement. I suggest we go
with a mix of this and state.